% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_training_log.R
\name{load_training_log}
\alias{load_training_log}
\title{Import training log from yaml}
\usage{
load_training_log(
  path = "inst/tinytest/test.yml",
  order = "dm",
  year = lubridate::year(Sys.time())
)
}
\arguments{
\item{order}{character same as in `?lubridate::parse_date_time()`}
}
\description{
Import training log from yaml
}
