% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swoleR-package.R
\docType{package}
\name{swoleR-package}
\alias{swoleR}
\alias{swoleR-package}
\title{swoleR: What the Package Does Using Title Case}
\description{
A package to help me get shredded and learn some sweet shinyr.
}
\keyword{internal}
